requirements.txt
langchain
streamlit
python-dotenv
langserve
fastapi
unicorn
sse_starlette
bs4
pypdf
chromadb
wikipedia
arxiv